Refactor to NlpExtractor
